requirements.txt
# Core streaming
kafka-python==2.0.2
pyspark==3.4.2

# Faker & validation
faker==24.2.0
pydantic==2.6.3

# AWS
boto3==1.34.51
aws-glue-sessions==1.0.5   # only needed on Glue dev endpoint

# Config
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.1